requirements.txt
streamlit==1.45.1
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1